move the table from Home to component/CarsTable.js
