+++ src/containers/Home.js
@@ -1,9 +1,9 @@
 import { useEffect } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
 import { createCarsList } from '../actions/index';
 import { carsListRequest, deleteCarRequest } from '../logic/carRequests';
+import CarsTable from '../components/CarsTable';
 
 const Home = ({ authentication, carsList, createCarsList }) => {
   useEffect(() => {
@@ -12,56 +12,14 @@ const Home = ({ authentication, carsList, createCarsList }) => {
     }
   }, []);
 
-  const deleteCar = (e) => {
-    const carId = parseInt(e.target.dataset.carId, 10);
+  const deleteCar = (carId) => {
     deleteCarRequest(authentication.authToken, carId);
   };
 
   let carsTable = null;
-  let deleteTH = null;
-  let editTH = null;
-
   if (carsList.length > 0) {
-    if (authentication.admin) {
-      deleteTH = (<th>delete the car</th>);
-      editTH = (<th>edit the car</th>);
-    }
     carsTable = (
-      <table>
-        <thead>
-          <tr>
-            <th>Mark</th>
-            <th>Model</th>
-            <th>Fabrication year</th>
-            <th>Price of trying</th>
-            <th>make appointment</th>
-            {deleteTH}
-            {editTH}
-          </tr>
-        </thead>
-        <tbody>
-          {carsList.map((e) => {
-            let deleteTD = null;
-            let editTD = null;
-            if (authentication.admin) {
-              deleteTD = (<td><button type="button" data-car-id={e.id} onClick={deleteCar}>delete the car</button></td>);
-              editTD = (<td><Link to={{ pathname: '/edit_car', state: { car: e } }}>edit the car</Link></td>);
-            }
-            return (
-              <tr key={`${e.mark} ${e.model}`}>
-                <td>{e.mark}</td>
-                <td>{e.model}</td>
-                <td>{e.year}</td>
-                <td>{e.price}</td>
-                <td><Link to={{ pathname: '/make-appointment', state: { car: e } }}>Make appointment</Link></td>
-                {deleteTD}
-                {editTD}
-              </tr>
-            );
-          })}
-        </tbody>
-      </table>
-
+      <CarsTable admin={authentication.admin} cars={carsList} clickHandler={deleteCar} />
     );
   }
